HMM/requirements.txt
sklearn
numpy
hmmlearn
Pillow
pickle